requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
mysql-connector-python
python-multipart
python-jose
passlib[bcrypt]
